// client/src/graphql/types/getAllTags.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getAllTags
// ====================================================

export interface getAllTags_tags {
  __typename: "Tag";
  id: number;
  tag: string;
}

export interface getAllTags {
  tags: getAllTags_tags[];
}